fixed minor issues
