+++ src/main/java/com/app/sellerretailreports/SellerRetailReportsApplication.java
@@ -2,12 +2,10 @@ package com.app.sellerretailreports;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableScheduling
-@EnableCaching
 public class SellerRetailReportsApplication {
     public static void main(String[] args) {
         SpringApplication.run(SellerRetailReportsApplication.class, args);
