+++ src/main/java/com/app/sellerretailreports/service/file/FileReaderService.java
@@ -8,26 +8,13 @@ import java.util.stream.Collectors;
 
 public class FileReaderService {
     public static String readFile(String fileName) {
-
-        try {
-            BufferedReader bufferedReader = Files.newBufferedReader(Paths.get(fileName));
-
-            try {
-                String str = bufferedReader.lines()
-                        .collect(Collectors.joining(System.lineSeparator()));
-                bufferedReader.close();
-                return str;
-            } catch (Throwable throwable) {
-                try {
-                    bufferedReader.close();
-                } catch (Throwable throwable1) {
-                    throwable.addSuppressed(throwable1);
-                }
-                throw throwable;
-            }
+        try (BufferedReader bufferedReader = Files.newBufferedReader(Paths.get(fileName))) {
+            String str = bufferedReader.lines()
+                    .collect(Collectors.joining(System.lineSeparator()));
+            bufferedReader.close();
+            return str;
         } catch (IOException e) {
             throw new RuntimeException("Cannot read data from the file " + fileName, e);
         }
-
     }
 }
