+++ src/main/resources/application.properties
@@ -8,3 +8,5 @@ caching.spring.TTL=420000
 
 jwt.expiration=${JWT_EXPIRATION}
 jwt.secret=${JWT_SECRET}
+
+logging.level.org.springframework.data.mongodb.core.MongoTemplate=DEBUG
